requirements.txt
absl-py==2.0.0
accelerate==0.21.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
blinker==1.7.0
branca==0.7.1
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.7
clickhouse-connect==0.7.0
coloredlogs==15.0.1
colorlog==6.8.0
comm==0.2.0
contourpy==1.1.1
cramjam==2.7.0
cryptography==42.0.0
cycler==0.12.0
dataclasses-json==0.5.14
datasets==2.16.1
DAWG-Python==0.7.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
docopt==0.6.2
duckdb==0.7.1
einops==0.6.1
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.0
fastjsonschema==2.18.0
fastparquet==2023.10.1
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
folium==0.15.1
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.5.4
gguf==0.6.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
ipykernel==6.27.1
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.15.0
kiwisolver==1.4.5
lab==7.4
langchain==0.0.267
langchain-community==0.0.16
langchain-core==0.1.16
langsmith==0.0.83
libclang==16.0.6
llama_cpp_python==0.2.28
lxml==4.9.4
lz4==4.3.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.8
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.16.3
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.5.0
orjson==3.9.7
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pdfminer.six==20221105
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.1.0
posthog==3.3.2
prometheus-client==0.17.1
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.0
pydotplus==2.0.2
Pygments==2.17.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.35.1
statsmodels==0.14.1
streamlit==1.25.0
streamlit-chat==0.1.1
sympy==1.12
tapi-wrapper2==0.1.3
tapi-yandex-metrika==2022.4.8
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.37.2
txt2tags==3.8
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.0.5
uvicorn==0.18.3
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xgboost==2.0.3
xxhash==3.4.1
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
